@@ -1,12 +1,10 @@
 package com.paymybuddy.paymybuddy.controller;
 
-import com.paymybuddy.paymybuddy.repository.UserRepository;
-import com.paymybuddy.paymybuddy.security.config.JwtTokenUtil;
 import com.paymybuddy.paymybuddy.security.model.request.JwtRequest;
-import com.paymybuddy.paymybuddy.security.service.LoggerService;
+import com.paymybuddy.paymybuddy.security.model.response.JwtResponse;
+import com.paymybuddy.paymybuddy.security.service.AuthenticationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,25 +13,17 @@ import org.springframework.web.bind.annotation.*;
 public class JwtAuthenticationController {
 
     @Autowired
-    private AuthenticationManager authenticationManager;
-
-    @Autowired
-    private JwtTokenUtil jwtTokenUtil;
-
-    @Autowired
-    private LoggerService userDetailsService;
-
-    @Autowired
-    UserRepository userRepository;
+    private AuthenticationService userDetailsService;
 
     @PostMapping(value = "/login")
-    public ResponseEntity<?> createAuthenticationToken(@RequestBody JwtRequest jwtRequest) throws Exception {
-        return userDetailsService.createAuthenticationToken(jwtRequest);
+    public ResponseEntity<JwtResponse> createAuthenticationToken(@RequestBody JwtRequest jwtRequest) {
+        JwtResponse authenticationToken = userDetailsService.createAuthenticationToken(jwtRequest);
+        return authenticationToken == null ? ResponseEntity.noContent().build() : ResponseEntity.ok(authenticationToken);
     }
 
 
     @PostMapping(value = "/register")
-    public ResponseEntity<?> Register(@RequestBody JwtRequest registerRequest) {
-        return userDetailsService.save(registerRequest);
+    public ResponseEntity<String> Register(@RequestBody JwtRequest registerRequest) {
+        return ResponseEntity.ok(userDetailsService.save(registerRequest));
     }
 }
