@@ -12,10 +12,8 @@ import org.apache.commons.validator.routines.EmailValidator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.DisabledException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -24,8 +22,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.transaction.Transactional;
 
-@Service
-public class LoggerService implements UserDetailsService {
+@Service()
+public class AuthenticationService implements UserDetailsService {
 
     @Autowired
     private UserRepository userRepository;
@@ -39,7 +37,7 @@ public class LoggerService implements UserDetailsService {
     @Autowired
     private AuthenticationManager authenticationManager;
 
-    private final Logger logger = LogManager.getLogger(LoggerService.class);
+    private final Logger logger = LogManager.getLogger(AuthenticationService.class);
 
     @Override
     @Transactional
@@ -52,15 +50,29 @@ public class LoggerService implements UserDetailsService {
         return UserDetailsImpl.build(user);
     }
 
-    public ResponseEntity<?> createAuthenticationToken(@RequestBody JwtRequest jwtRequest) throws Exception {
+    public JwtResponse createAuthenticationToken(@RequestBody JwtRequest jwtRequest) {
         checkValidMail(jwtRequest.getEmail());
         if (authenticate(jwtRequest.getEmail(), jwtRequest.getPassword())) {
             final UserDetailsImpl userDetails = loadUserByUsername(jwtRequest.getEmail());
             final String token = jwtTokenUtil.generateToken(userDetails);
             logger.info("Request login successful");
-            return ResponseEntity.ok(new JwtResponse(token, userDetails.getUsername(), userDetails.getId()));
+            return new JwtResponse(token, userDetails.getUsername(), userDetails.getId());
         }
-        return ResponseEntity.notFound().build();
+        return null;
+    }
+
+    @Transactional
+    public String save(JwtRequest registerRequest) {
+        checkValidMail(registerRequest.getEmail());
+        if (userRepository.existsByEmail(registerRequest.getEmail())) {
+            throw new NoUserFoundException(registerRequest.getEmail());
+        }
+        User user = new User(registerRequest.getEmail(),
+                webSecurityConfig.passwordEncoder().encode(registerRequest.getPassword()));
+
+        userRepository.save(user);
+        logger.info("Request register successful");
+        return "User registered successfully!";
     }
 
     private boolean authenticate(String username, String password) {
@@ -72,23 +84,10 @@ public class LoggerService implements UserDetailsService {
         return true;
     }
 
-    @Transactional
-    public ResponseEntity save(JwtRequest registerRequest) {
-        checkValidMail(registerRequest.getEmail());
-        if (userRepository.existsByEmail(registerRequest.getEmail())) {
-            throw new NoUserFoundException(registerRequest.getEmail());
-        }
-        User user = new User(registerRequest.getEmail(),
-                webSecurityConfig.passwordEncoder().encode(registerRequest.getPassword()));
-
-        userRepository.save(user);
-        logger.info("Request register successful");
-        return ResponseEntity.ok("User registered successfully!");
-    }
-
     private void checkValidMail(String email) {
         if (!EmailValidator.getInstance().isValid(email)) {
             throw new NonValidEmailLogin(email);
         }
     }
 }
+
