@@ -1,14 +1,15 @@
 package com.paymybuddy.paymybuddy.service;
 
+import com.paymybuddy.paymybuddy.data.TestData;
+import com.paymybuddy.paymybuddy.exception.ContactAlreadyAssignedException;
 import com.paymybuddy.paymybuddy.model.Contact;
 import com.paymybuddy.paymybuddy.repository.ContactRepository;
 import com.paymybuddy.paymybuddy.repository.UserRepository;
-import com.paymybuddy.paymybuddy.service.data.TestData;
 import com.paymybuddy.paymybuddy.service.impl.ContactServiceImpl;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -47,5 +48,12 @@ public class ContactServiceTest {
         Mockito.when(contactRepository.findByUserIdAndContactId(Mockito.anyLong(), Mockito.anyLong())).thenReturn(null);
         contactService.addContact("clement", SecurityContextHolder.getContext().getAuthentication());
         Mockito.verify(contactRepository, Mockito.times(1)).save(Mockito.any());
+    }
+
+    @Test
+    public void contactAlreadyAssignedTest() {
+        Mockito.when(contactRepository.findByUserIdAndContactId(Mockito.anyLong(), Mockito.anyLong())).thenReturn(new Contact());
+        Assertions.assertThatThrownBy(() -> contactService.addContact("test@gmail.com",  SecurityContextHolder.getContext().getAuthentication())
+        ).isInstanceOf(ContactAlreadyAssignedException.class);
     }
 }
