@@ -1,16 +1,20 @@
 package com.paymybuddy.paymybuddy.service;
 
+import com.paymybuddy.paymybuddy.data.TestData;
+import com.paymybuddy.paymybuddy.exception.NoBankFoundException;
+import com.paymybuddy.paymybuddy.exception.NoEnoughMoneyOnBalanceException;
+import com.paymybuddy.paymybuddy.exception.NonValidAmountException;
 import com.paymybuddy.paymybuddy.model.Bank;
 import com.paymybuddy.paymybuddy.model.User;
 import com.paymybuddy.paymybuddy.repository.BankRepository;
 import com.paymybuddy.paymybuddy.repository.OperationRepository;
 import com.paymybuddy.paymybuddy.repository.UserRepository;
-import com.paymybuddy.paymybuddy.service.data.TestData;
 import com.paymybuddy.paymybuddy.service.impl.TransferServiceImpl;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -75,5 +79,25 @@ public class TransferServiceTest {
         Assertions.assertThat(transferService.transferMoneyToUser(
                 "BNK_BNP", 100.00, SecurityContextHolder.getContext().getAuthentication())
         ).isEqualTo(TestData.getUserDTOTransferMoneyToBankOrUser());
+    }
+
+    @Test
+    public void NonValidAmountExceptionTest() {
+        Assertions.assertThatThrownBy(() -> transferService.transferMoneyToUser("test", 300.222,  SecurityContextHolder.getContext().getAuthentication())
+        ).isInstanceOf(NonValidAmountException.class);
+    }
+
+    @Test
+    public void NoEnoughMoneyOnBalanceExceptionTest() {
+        Mockito.when(userRepository.findBalanceById(Mockito.anyLong())).thenReturn(10.00);
+        Assertions.assertThatThrownBy(() -> transferService.transferMoneyToUser("test", 11.00,  SecurityContextHolder.getContext().getAuthentication())
+        ).isInstanceOf(NoEnoughMoneyOnBalanceException.class);
+    }
+
+    @Test
+    public void NoBankFoundException() {
+        Mockito.when(bankRepository.findByName(Mockito.anyString())).thenReturn(null);
+        Assertions.assertThatThrownBy(() -> transferService.transferMoneyToBank("BNP", 10.00,  SecurityContextHolder.getContext().getAuthentication())
+        ).isInstanceOf(NoBankFoundException.class);
     }
 }
